// sanity/sanity.types.ts
export type Workout = {
  _id: string
  _type: 'workout'
  _createdAt: string
  _updatedAt: string
  _rev: string
  userId?: string
  date?: string
  exercises?: Array<{
    exercise?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'exercise'
    }
    duration?: number
    sets?: Array<{
      reps?: number
      weightType?: 'kg' | 'lbs'
      weightValue?: number
      _key: string
    }>
    _type: 'exerciseLog'
    _key: string
  }>
}
// ...
export declare const internalGroqTypeReferenceTo: unique symbol